requirements.txt
httpx
telethon
python_dotenv
matplotlib
plotly
bokeh
cython
setuptools
xlrd==1.2.0
folium
scikit-learn
numpy

